scripts/requirements.txt
# Python dependencies for image pair generation script

# Google Generative AI SDK (for Gemini and Imagen)
google-genai>=0.2.0

# Supabase Python client
supabase>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Image processing
Pillow>=10.0.0

